chore: extract state helpers into sub-module
